fix: migrated views store to pinia
