+++ src/store/index.js
@@ -3,7 +3,6 @@ import Vuex from 'vuex';
 import activity from './modules/activity';
 import buckets from './modules/buckets';
 import categories from './modules/categories';
-import views from './modules/views';
 //import createLogger from '../../../src/plugins/logger';
 
 Vue.use(Vuex);
@@ -15,7 +14,6 @@ export default new Vuex.Store({
     activity,
     buckets,
     categories,
-    views,
   },
   strict: debug,
   //  plugins: debug ? [createLogger()] : [],
